Crate chill

Reexports

pub use path::{DatabaseName, DatabaseNameRef, DatabasePath, DatabasePathRef, DesignDocumentName, DesignDocumentNameRef, DocumentId, DocumentIdRef, DocumentPath, DocumentPathRef, IntoDatabasePath, IntoDocumentPath, LocalDocumentName, LocalDocumentNameRef, NormalDocumentName, NormalDocumentNameRef};

Modules

action

path
Types and traits for specifying databases, documents, views, etc.

testing
Utilities for testing a CouchDB application.

Structs

Client

Document

ErrorResponse
Error information returned from the CouchDB server when an error occurs while processing the client's request.

Revision
A document revision, which uniquely identifies a version of a document.

Enums

Error

Traits

IntoUrl
Types that are convertible into a URL.