Enum chill::Error

pub enum Error {
    DatabaseExists(ErrorResponse),
    DocumentConflict(ErrorResponse),
    NotFound(ErrorResponse),
    Unauthorized(ErrorResponse),
    // some variants omitted
}

Contains information for an error originating from or propagated by Chill.

Variants
DatabaseExists(ErrorResponse)

The database already exists.

DocumentConflict(ErrorResponse)

A document with the same id already exists or the given revision is not the latest revision for the document.

NotFound(ErrorResponse)

The target resource—e.g., database, document, etc.—does not exist or is deleted.

Unauthorized(ErrorResponse)

The client lacks permission to complete the action.