Enum chill::ViewResponse

pub enum ViewResponse<K: Deserialize, V: Deserialize> {
    Reduced(ReducedView<V>),
    Unreduced(UnreducedView<K, V>),
}

Contains the response from the CouchDB server as a result of successfully executing a view.

A ViewResponse takes one of two forms, reduced or unreduced, depending on whether the view's reduce function ran. See the CouchDB documentation for more details.

Although ViewResponse implements the Ord and PartialOrd traits, Chill makes no guarantee how that ordering is defined and may change the definition in an upcoming release. Chill defines the ordering only so that applications may use ViewResponse in ordered collections such as trees.

Variants

Reduced

Contains a reduced view response.

Unreduced

Contains an unreduced view response.

Methods

impl<K: Deserialize, V: Deserialize> ViewResponse<K, V>

fn as_reduced(&self) -> Option<&ReducedView<V>>

Returns the view response in its reduced form, if the response is reduced.

fn as_reduced_mut(&mut self) -> Option<&mut ReducedView<V>>

Returns the view response in its reduced form, if the response is reduced.

fn as_unreduced(&self) -> Option<&UnreducedView<K, V>>

Returns the view response in its unreduced form, if the response is unreduced.

fn as_unreduced_mut(&mut self) -> Option<&mut UnreducedView<K, V>>

Returns the view response in its unreduced form, if the response is unreduced.