Struct chill::Revision

pub struct Revision {
    // some fields omitted
}

A document revision, which uniquely identifies a version of a document.

A document revision comprises a sequence number and an MD5 digest. The sequence number (usually) starts at 1 when the document is created and increments by one each time the document is updated. The digest is a hash of the document content.

In serialized form, a revision looks like 1-9c65296036141e575d32ba9c034dd3ee.

Methods

impl Revision

fn parse(s: &str) -> Result<Self, Error>

Constructs a new Revision from the given string.

The string must be of the form 42-1234567890abcdef1234567890abcdef.

fn sequence_number(&self) -> u64

Returns the sequence number part of the revision.

The sequence number is the 123 part of the revision 123-00000000000000000000000000000000.