Struct chill::UnreducedView

pub struct UnreducedView<K: Deserialize, V: Deserialize> {
    // some fields omitted
}

Contains a view response in unreduced form.

Methods

impl<K: Deserialize, V: Deserialize> UnreducedView<K, V>

fn total_rows(&self) -> u64

Returns the number of all rows in the view response, including rows excluded from the vector.

fn offset(&self) -> u64

Returns the number of rows in the view response excluded before the first row in the vector.

fn update_sequence_number(&self) -> Option<u64>

Returns the update sequence number the view reflects, if available.

fn rows(&self) -> &Vec<ViewRow<K, V>>

Returns the vector containing all rows in the view response.