Struct chill::ViewFunction

pub struct ViewFunction {
    pub map: String,
    pub reduce: Option<String>,
    // some fields omitted
}

Container for a map and optional reduce function of a view.

ViewFunction is a convenience type for applications that work with view functions in design documents. For more information about view functions and design documents, please see the CouchDB documentation.

Examples

extern crate chill;

let view_function = chill::ViewFunction::new_with_reduce(
    "function(doc) { emit(doc.key_thing, doc.value_thing); }",
    "_sum");

assert_eq!("function(doc) { emit(doc.key_thing, doc.value_thing); }",
           view_function.map);
assert_eq!(Some(String::from("_sum")), view_function.reduce);

Fields
map: String

The view's map function.

For more information about map functions, please see the CouchDB documentation.

reduce: Option<String>

The view's reduce function, if available.

For more information about reduce functions, please see the CouchDB documentation.

Methods

impl ViewFunction

fn new<M: Into<String>>(map: M) -> Self

Constructs a new ViewFunction that has no reduce function.

fn new_with_reduce<M: Into<String>, R: Into<String>>(map: M, reduce: R) -> Self

Constructs a new ViewFunction that has a reduce function.