Struct chill::ViewResponse

pub struct ViewResponse {
    // some fields omitted
}

Methods

impl ViewResponse

fn total_rows(&self) -> Option<u64>

Returns how many rows are in the view, including rows excluded in the response, if available.

The total number of rows is available if and only if the view is unreduced. Even group-reduced views, which may contain multiple rows in the response, have no total number of rows associated with them.

fn offset(&self) -> Option<u64>

Returns how many rows are excluded from the view response that are ordered before the first row in the response, if available.

The offset is available if and only if the view is unreduced.

fn update_sequence_number(&self) -> Option<u64>

Returns the update sequence number that the view reflects, if available.

fn rows(&self) -> &Vec<ViewRow>

Returns the vector containing all rows in the view response.