Struct chill::ViewRow

pub struct ViewRow<K: Deserialize, V: Deserialize> {
    // some fields omitted
}

Contains a single row in an unreduced view response.

See the CouchDB documentation for more details about view rows.

Although ViewRow implements the Ord and PartialOrd traits, Chill makes no guarantee how that ordering is defined and may change the definition in an upcoming release. One consequence is that if an application sorts view rows itself then the result may be in a different order than if the CouchDB server had sorted the rows. This is an anti-pattern; applications should rely on the server to do sorting. Chill defines the ordering only so that applications may use ViewRow in ordered collections such as trees.

Methods

impl<K, V> ViewRow<K, V> where K: Deserialize, V: Deserialize

fn key(&self) -> &K

Returns the row's key.

fn value(&self) -> &V

Returns the row's value.

fn document_path(&self) -> DocumentPathRef

Returns the path of the row's document.