Struct chill::testing::ViewResponseBuilder

pub struct ViewResponseBuilder<K: Deserialize, V: Deserialize, M> {
    // some fields omitted
}

Builds a view response.

A ViewResponseBuilder constructs a view response as though the response originated from a CouchDB Server. This allows an application to mock a view response without using a production database.

Examples

One point of inconvenience when using a ViewResponseBuilder is that type inference doesn't work when building an unreduced view response, so applications should use the turbofish (::<Key, Value, _>) to specify the key and value types, like so:

use chill::testing::ViewResponseBuilder;

ViewResponseBuilder::<String, i32, _>::new_unreduced(123,       // `total_rows`
                                                     0,         // `offset`
                                                     "baseball" // database name
                                                    )
    .with_row("Hammerin' Hank", 755, "hank_aaron_doc_id")
    .with_row("The Bambino", 714, "babe_ruth_doc_id")
    .unwrap();

Reduced views are easier because type inference does work.

use chill::testing::ViewResponseBuilder;

// The view response key is inferred as () because reduced views have no
// key.
ViewResponseBuilder::new_reduced(1469).unwrap();

Methods

impl<V: Deserialize> ViewResponseBuilder<(), V, ViewIsReduced>

fn new_reduced(value: V) -> Self

Constructs a reduced view response.

fn with_update_sequence_number(self, update_seq: u64) -> Self

Sets the update sequence number for the view response.

The update sequence number corresponds to the update_seq field in the view response. By default, its value is None.

fn unwrap(self) -> ViewResponse<(), V>

Returns the builder's view response.

impl<K: Deserialize, V: Deserialize> ViewResponseBuilder<K, V, ViewIsUnreduced>

fn new_unreduced<D: Into<DatabaseName>>(total_rows: u64, offset: u64, db_name: D) -> Self

Constructs an unreduced view response.

Any rows added via the with_row method will use the given database name as part of their document path.

fn with_row<D, IntoK, IntoV>(self, key: IntoK, value: IntoV, doc_id: D) -> Self where D: Into<DocumentId>, IntoK: Into<K>, IntoV: Into<V>

Appends a new row into the view response.

fn with_update_sequence_number(self, update_seq: u64) -> Self

Sets the update sequence number for the view response.

The update sequence number corresponds to the update_seq field in the view response. By default, its value is None.

fn unwrap(self) -> ViewResponse<K, V>

Returns the builder's view response.